AutomationTool/requirements.txt
ldap3
pyyaml
impacket
